# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Reliable Fresh Export Management System",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/pre-cooling": {
      "post": {
        "tags": [
          "pre-cooling"
        ],
        "summary": "Create Pre Cooling",
        "operationId": "create_pre_cooling_api_v1_pre_cooling_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreCoolingCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PreCoolingRead"
                  },
                  "title": "Response Create Pre Cooling Api V1 Pre Cooling Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "pre-cooling"
        ],
        "summary": "List Pre Cooling",
        "operationId": "list_pre_cooling_api_v1_pre_cooling_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "incomplete_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Incomplete Only"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PreCoolingRead"
                  },
                  "title": "Response List Pre Cooling Api V1 Pre Cooling Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pre-cooling/{record_id}/complete": {
      "patch": {
        "tags": [
          "pre-cooling"
        ],
        "summary": "Complete Pre Cooling",
        "operationId": "complete_pre_cooling_api_v1_pre_cooling__record_id__complete_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Record Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreCoolingComplete"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreCoolingRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/purchase-orders": {
      "post": {
        "tags": [
          "purchase orders"
        ],
        "summary": "Create Po",
        "operationId": "create_po_api_v1_purchase_orders_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/POCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PORead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "purchase orders"
        ],
        "summary": "List Pos",
        "operationId": "list_pos_api_v1_purchase_orders_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/POStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PORead"
                  },
                  "title": "Response List Pos Api V1 Purchase Orders Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/purchase-orders/{po_id}": {
      "get": {
        "tags": [
          "purchase orders"
        ],
        "summary": "Get Po",
        "operationId": "get_po_api_v1_purchase_orders__po_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "po_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Po Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PORead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/purchase-orders/{po_id}/status": {
      "patch": {
        "tags": [
          "purchase orders"
        ],
        "summary": "Update Po Status",
        "operationId": "update_po_status_api_v1_purchase_orders__po_id__status_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "po_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Po Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/POStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PORead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "POCreate": {
        "properties": {
          "po_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Po Date"
          },
          "supplier_name": {
            "type": "string",
            "minLength": 1,
            "title": "Supplier Name"
          },
          "supplier_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Address"
          },
          "supplier_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Email"
          },
          "supplier_gst": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Gst"
          },
          "payment_terms": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Terms"
          },
          "supplier_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Ref"
          },
          "other_refs": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Other Refs"
          },
          "dispatch_through": {
            "type": "string",
            "title": "Dispatch Through",
            "default": "Road Transport"
          },
          "destination": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination"
          },
          "freight": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Freight"
          },
          "other_charges": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Other Charges",
            "default": "0"
          },
          "line_items": {
            "items": {
              "$ref": "#/components/schemas/POLineItemCreate"
            },
            "type": "array",
            "minItems": 1,
            "title": "Line Items"
          }
        },
        "type": "object",
        "required": [
          "supplier_name",
          "line_items"
        ],
        "title": "POCreate"
      },
...
      "PORead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "po_number": {
            "type": "string",
            "title": "Po Number"
          },
          "po_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Po Date"
          },
          "supplier_name": {
            "type": "string",
            "title": "Supplier Name"
          },
          "supplier_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Address"
          },
          "supplier_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Email"
          },
          "supplier_gst": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Gst"
          },
          "payment_terms": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Terms"
          },
          "supplier_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Ref"
          },
          "other_refs": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Other Refs"
          },
          "dispatch_through": {
            "type": "string",
            "title": "Dispatch Through"
          },
          "destination": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination"
          },
          "assessable_value": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assessable Value"
          },
          "cgst_total": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cgst Total"
          },
          "sgst_total": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sgst Total"
          },
          "freight": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Freight"
          },
          "other_charges": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Other Charges"
          },
          "grand_total": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grand Total"
          },
          "total_in_words": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total In Words"
          },
          "status": {
            "$ref": "#/components/schemas/POStatus"
          },
          "created_by": {
            "type": "integer",
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "line_items": {
            "items": {
              "$ref": "#/components/schemas/POLineItemRead"
            },
            "type": "array",
            "title": "Line Items",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "po_number",
          "po_date",
          "supplier_name",
          "supplier_address",
          "supplier_email",
          "supplier_gst",
          "payment_terms",
          "supplier_ref",
          "other_refs",
          "dispatch_through",
          "destination",
          "assessable_value",
          "cgst_total",
          "sgst_total",
          "freight",
          "other_charges",
          "grand_total",
          "total_in_words",
          "status",
          "created_by",
          "created_at"
        ],
        "title": "PORead"
      },
      "POStatus": {
        "type": "string",
        "enum": [
          "draft",
          "issued",
          "completed"
        ],
        "title": "POStatus"
      },
      "POStatusUpdate": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/POStatus"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "POStatusUpdate"
      },
...
      "PreCoolingComplete": {
        "properties": {
          "out_time": {
            "type": "string",
            "format": "time",
            "title": "Out Time"
          },
          "out_berry_temp": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Out Berry Temp"
          }
        },
        "type": "object",
        "required": [
          "out_time",
          "out_berry_temp"
        ],
        "title": "PreCoolingComplete"
      },
      "PreCoolingCreate": {
        "properties": {
          "pallet_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "minItems": 1,
            "title": "Pallet Ids"
          },
          "date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "num_boxes": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Num Boxes"
          },
          "in_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "time"
              },
              {
                "type": "null"
              }
            ],
            "title": "In Time"
          },
          "in_berry_temp": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "In Berry Temp"
          },
          "out_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Out Time"
          },
          "out_berry_temp": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Out Berry Temp"
          }
        },
        "type": "object",
        "required": [
          "pallet_ids"
        ],
        "title": "PreCoolingCreate",
        "description": "Batch entry: one record is created per pallet_id (frontend prompt \u00a718).\nOut-fields omitted = partial save; complete later via PATCH."
      },
      "PreCoolingRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "pallet_id": {
            "type": "integer",
            "title": "Pallet Id"
          },
          "date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "num_pallets": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Num Pallets"
          },
          "num_boxes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Num Boxes"
          },
          "in_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "time"
              },
              {
                "type": "null"
              }
            ],
            "title": "In Time"
          },
          "in_berry_temp": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "In Berry Temp"
          },
          "out_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Out Time"
          },
          "out_berry_temp": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Out Berry Temp"
          },
          "created_by": {
            "type": "integer",
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "is_complete": {
            "type": "boolean",
            "title": "Is Complete",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "pallet_id",
          "date",
          "num_pallets",
          "num_boxes",
          "in_time",
          "in_berry_temp",
          "out_time",
          "out_berry_temp",
          "created_by",
          "created_at"
        ],
        "title": "PreCoolingRead"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}